requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --extra=dev --output-file=requirements/dev.txt pyproject.toml
#
babel==2.17.0
    # via mkdocs-material
backrefs==5.9
    # via mkdocs-material
black==25.9.0
    # via CodeCraftAI (pyproject.toml)
build==1.3.0
    # via pip-tools
certifi==2025.8.3
    # via requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   black
    #   mkdocs
    #   pip-tools
colorama==0.4.6
    # via
    #   griffe
    #   mkdocs-material
distlib==0.4.0
    # via virtualenv
filelock==3.20.0
    # via virtualenv
ghp-import==2.1.0
    # via mkdocs
griffe==1.14.0
    # via mkdocstrings-python
identify==2.6.15
    # via pre-commit
idna==3.11
    # via requests
iniconfig==2.3.0
    # via pytest
jinja2==3.1.6
    # via
    #   mkdocs
    #   mkdocs-material
    #   mkdocstrings
markdown==3.9
    # via
    #   mkdocs
    #   mkdocs-autorefs
    #   mkdocs-material
    #   mkdocstrings
    #   pymdown-extensions
markupsafe==3.0.3
    # via
    #   jinja2
    #   mkdocs
    #   mkdocs-autorefs
    #   mkdocstrings
mergedeep==1.3.4
    # via
    #   mkdocs
    #   mkdocs-get-deps
mkdocs==1.6.1
    # via
    #   CodeCraftAI (pyproject.toml)
    #   mkdocs-autorefs
    #   mkdocs-material
    #   mkdocstrings
mkdocs-autorefs==1.4.3
    # via
    #   mkdocstrings
    #   mkdocstrings-python
mkdocs-get-deps==0.2.0
    # via mkdocs
mkdocs-material==9.6.23
    # via CodeCraftAI (pyproject.toml)
mkdocs-material-extensions==1.3.1
    # via mkdocs-material
mkdocstrings[python]==0.30.1
    # via
    #   CodeCraftAI (pyproject.toml)
    #   mkdocstrings-python
mkdocstrings-python==1.18.2
    # via mkdocstrings
mypy-extensions==1.1.0
    # via black
nodeenv==1.9.1
    # via pre-commit
packaging==25.0
    # via
    #   black
    #   build
    #   mkdocs
    #   pytest
paginate==0.5.7
    # via mkdocs-material
pathspec==0.12.1
    # via
    #   black
    #   mkdocs
pip-tools==7.5.1
    # via CodeCraftAI (pyproject.toml)
platformdirs==4.5.0
    # via
    #   black
    #   mkdocs-get-deps
    #   virtualenv
pluggy==1.6.0
    # via pytest
pre-commit==4.3.0
    # via CodeCraftAI (pyproject.toml)
pygments==2.19.2
    # via
    #   mkdocs-material
    #   pytest
pymdown-extensions==10.16.1
    # via
    #   mkdocs-material
    #   mkdocstrings
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.4.2
    # via CodeCraftAI (pyproject.toml)
python-dateutil==2.9.0.post0
    # via ghp-import
pytokens==0.2.0
    # via black
pyyaml==6.0.3
    # via
    #   mkdocs
    #   mkdocs-get-deps
    #   pre-commit
    #   pymdown-extensions
    #   pyyaml-env-tag
pyyaml-env-tag==1.1
    # via mkdocs
requests==2.32.5
    # via mkdocs-material
ruff==0.14.3
    # via CodeCraftAI (pyproject.toml)
six==1.17.0
    # via python-dateutil
urllib3==2.5.0
    # via requests
virtualenv==20.35.4
    # via pre-commit
watchdog==6.0.0
    # via mkdocs
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
